+++ src/resources/transactions/operations/index.ts
@@ -13,3 +13,4 @@ export * from './update-transaction-request-body.js';
 export * from './transaction-preview-request-body.js';
 export * from './get-transaction-invoice-pdf-query-parameters.js';
 export * from './revise-transaction-request-body.js';
+export * from './non-catalog-discount.js';