+++ CHANGELOG.md
@@ -27,6 +27,12 @@ adding additional type guards.
 - Added support for new payment methods `blik`, `mb_way`, `pix` and `upi`. See [related changelog](https://developer.paddle.com/changelog/2025/blik-mbway-payment-methods?utm_source=dx&utm_medium=paddle-node-sdk).
 - Non-catalog discounts on Transactions, see [changelog](https://developer.paddle.com/changelog/2025/custom-discounts?utm_source=dx&utm_medium=paddle-node-sdk)
 
+### Changed
+
+- `CustomData` type definition is now `Record<string, any>`
+
+---
+
 ## 3.2.1 - 2025-08-26
 
 ### Fixed